Implement loading cards

+++ eventPlanner/src/main/java/com/ftn/iss/eventPlanner/model/specification/ProductSpecification.java
@@ -33,7 +33,7 @@ public class ProductSpecification {
                         : criteriaBuilder.conjunction();
     }
 
-    public static Specification<Product> betweenPrices(Integer minPrice, Integer maxPrice) {
+    public static Specification<Product> betweenPrices(Double minPrice, Double maxPrice) {
         return (root, query, criteriaBuilder) -> {
             if (minPrice == null && maxPrice == null) return criteriaBuilder.conjunction();
             if (minPrice != null && maxPrice != null) {
